Create README.md
The project involves building a machine learning model to predict the likelihood of an individual having diabetes based on specific medical and demographic data. The model is trained using a dataset that includes various health metrics and is then deployed using Streamlit for an interactive web-based interface.

Project Description: Diabetes Prediction Using Machine Learning
Project Overview
The project involves building a machine learning model to predict the likelihood of an individual having diabetes based on specific medical and demographic data. The model is trained using a dataset that includes various health metrics and is then deployed using Streamlit for an interactive web-based interface.

Objective
The primary objective of this project is to develop an accurate and reliable prediction system that can help in the early detection of diabetes. Early detection is crucial as it can lead to timely treatment and lifestyle changes, potentially preventing complications associated with diabetes